🧪test(e2e-test/crdt): playwright에  해당하는 부분 e2e-test로 이동
