+++ e2e-test/crdt/test/drawing-performance/event-data-1.json
