+++ e2e-test/crdt/test/drawing-performance/event-data-2.json
