+++ e2e-test/crdt/test/drawing-text/event-data-1.json
