+++ e2e-test/crdt/test/drawing-text/event-data-3.json
